+++ crates/guest-storage-apply/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [0.24.21](https://github.com/okou-ai/okou/compare/guest-storage-apply-v0.24.20...guest-storage-apply-v0.24.21) (2026-09-25)
+
 ## [0.24.20](https://github.com/okou-ai/okou/compare/guest-storage-apply-v0.24.19...guest-storage-apply-v0.24.20) (2026-09-24)
 
 ## [0.24.19](https://github.com/okou-ai/okou/compare/guest-storage-apply-v0.24.18...guest-storage-apply-v0.24.19) (2026-09-24)